chore: move component to the root and rename it to components
