+++ components/ClassCard.tsx
