+++ components/SemCard.tsx
